+++ DesktopEdge/MainWindow.xaml.cs
@@ -106,7 +106,10 @@ namespace ZitiDesktopEdge {
 				LogManager.Configuration = config;
 			}
 			logger.Info("============================== UI started ==============================");
-			logger.Info("logger initialized using file: {0}", byFile);
+			logger.Info("logger initialized");
+			logger.Info("    - version   : {0}", asm.GetName().Version.ToString());
+			logger.Info("    - using file: {0}", byFile);
+			logger.Info("========================================================================");
 
 			App.Current.MainWindow.WindowState = WindowState.Normal;
 			App.Current.MainWindow.Closing += MainWindow_Closing;