+++ ZitiUpdateService/Program.cs
@@ -44,8 +44,11 @@ namespace ZitiUpdateService {
 				// Apply config           
 				LogManager.Configuration = config;
 			}
-			Logger.Info("============================== ziti-monitor started ==============================");
-			Logger.Info("logger initialized using file: {0}", byFile);
+			Logger.Info("========================= ziti-monitor started =========================");
+			Logger.Info("logger initialized");
+			Logger.Info("    - version   : {0}", asm.GetName().Version.ToString());
+			Logger.Info("    - using file: {0}", byFile);
+			Logger.Info("========================================================================");
 
 			UpdateService updateSvc = new UpdateService();
 			updateSvc.AutoLog = true;